services/user-service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
pymemcache==4.0.0
pymongo==4.6.1
boto3==1.34.40
gunicorn==21.2.0
requests==2.31.0
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
